refactor(kda): adjust file names
