+++ libs/coin-modules/coin-kadena/src/__tests__/unit/estimateMaxSpendable.unit.test.ts
@@ -1,5 +1,5 @@
 import { fetchCoinDetailsForAccount } from "../../api/network";
-import estimateMaxSpendable from "../../js-estimateMaxSpendable";
+import estimateMaxSpendable from "../../estimateMaxSpendable";
 import { Transaction } from "../../types";
 import {
   account,
