+++ libs/coin-modules/coin-kadena/src/prepareTransaction.ts
