+++ libs/coin-modules/coin-kadena/src/synchronisation.ts
